+++ lib/text/String.dart
@@ -1,9 +1,8 @@
 part of rocket_x;
 
 extension StringExtension on String {
-
+  /// Transforms a [String] widget to a [Text] widget
   Text get xToText {
     return Text(this);
   }
-
 }