+++ lib/text/TextStyle.dart
@@ -33,11 +33,14 @@ extension TextStyleExtension on Text {
     );
   }
 
+  /// Sets the TextColor property
   Text xTextColor(Color color) {
     return Text(
       this.data,
       key: this.key,
-      style: this.style != null ? this.style.copyWith(color: color) : TextStyle(color: color),
+      style: this.style != null
+          ? this.style.copyWith(color: color)
+          : TextStyle(color: color),
       locale: this.locale,
       semanticsLabel: this.semanticsLabel,
       textWidthBasis: this.textWidthBasis,
@@ -52,11 +55,14 @@ extension TextStyleExtension on Text {
     );
   }
 
+  /// Sets the FontFamily property
   Text xFontFamily(String family) {
     return Text(
       this.data,
       key: this.key,
-      style: this.style != null ? this.style.copyWith(fontFamily: family) : TextStyle(fontFamily: family),
+      style: this.style != null
+          ? this.style.copyWith(fontFamily: family)
+          : TextStyle(fontFamily: family),
       locale: this.locale,
       semanticsLabel: this.semanticsLabel,
       textWidthBasis: this.textWidthBasis,
@@ -71,12 +77,15 @@ extension TextStyleExtension on Text {
     );
   }
 
+  /// Sets the WordSpacing property
   Text xWordSpacing(double space) {
     assert(space >= 0);
     return Text(
       this.data,
       key: this.key,
-      style: this.style != null ? this.style.copyWith(wordSpacing: space) : TextStyle(wordSpacing: space),
+      style: this.style != null
+          ? this.style.copyWith(wordSpacing: space)
+          : TextStyle(wordSpacing: space),
       locale: this.locale,
       semanticsLabel: this.semanticsLabel,
       textWidthBasis: this.textWidthBasis,
@@ -91,11 +100,14 @@ extension TextStyleExtension on Text {
     );
   }
 
+  /// Sets the FontSize property
   Text xFontSize(double size) {
     return Text(
       this.data,
       key: this.key,
-      style: this.style != null ? this.style.copyWith(fontSize: size) : TextStyle(fontSize: size),
+      style: this.style != null
+          ? this.style.copyWith(fontSize: size)
+          : TextStyle(fontSize: size),
       locale: this.locale,
       semanticsLabel: this.semanticsLabel,
       textWidthBasis: this.textWidthBasis,
@@ -110,11 +122,14 @@ extension TextStyleExtension on Text {
     );
   }
 
+  /// Sets the BackgroundColor property
   Text xBackgroundColor(Color color) {
     return Text(
       this.data,
       key: this.key,
-      style: this.style != null ? this.style.copyWith(backgroundColor: color) : TextStyle(backgroundColor: color),
+      style: this.style != null
+          ? this.style.copyWith(backgroundColor: color)
+          : TextStyle(backgroundColor: color),
       locale: this.locale,
       semanticsLabel: this.semanticsLabel,
       textWidthBasis: this.textWidthBasis,
@@ -129,11 +144,14 @@ extension TextStyleExtension on Text {
     );
   }
 
+  /// Sets the Background property
   Text xBackground(Paint background) {
     return Text(
       this.data,
       key: this.key,
-      style: this.style != null ? this.style.copyWith(background: background) : TextStyle(background: background),
+      style: this.style != null
+          ? this.style.copyWith(background: background)
+          : TextStyle(background: background),
       locale: this.locale,
       semanticsLabel: this.semanticsLabel,
       textWidthBasis: this.textWidthBasis,
@@ -148,11 +166,14 @@ extension TextStyleExtension on Text {
     );
   }
 
+  /// Sets the Shadows property
   Text xShadows(List<Shadow> shadow) {
     return Text(
       this.data,
       key: this.key,
-      style: this.style != null ? this.style.copyWith(shadows: shadow) : TextStyle(shadows: shadow),
+      style: this.style != null
+          ? this.style.copyWith(shadows: shadow)
+          : TextStyle(shadows: shadow),
       locale: this.locale,
       semanticsLabel: this.semanticsLabel,
       textWidthBasis: this.textWidthBasis,
@@ -167,11 +188,14 @@ extension TextStyleExtension on Text {
     );
   }
 
+  /// Sets the FontStyle property
   Text xFontStyle(FontStyle fstyle) {
     return Text(
       this.data,
       key: this.key,
-      style: this.style != null ? this.style.copyWith(fontStyle: fstyle) : TextStyle(fontStyle: fstyle),
+      style: this.style != null
+          ? this.style.copyWith(fontStyle: fstyle)
+          : TextStyle(fontStyle: fstyle),
       locale: this.locale,
       semanticsLabel: this.semanticsLabel,
       textWidthBasis: this.textWidthBasis,
@@ -186,11 +210,14 @@ extension TextStyleExtension on Text {
     );
   }
 
+  /// Sets the FontWeight property
   Text xFontWeight(FontWeight fontWeight) {
     return Text(
       this.data,
       key: this.key,
-      style: this.style != null ? this.style.copyWith(fontWeight: fontWeight) : TextStyle(fontWeight: fontWeight),
+      style: this.style != null
+          ? this.style.copyWith(fontWeight: fontWeight)
+          : TextStyle(fontWeight: fontWeight),
       locale: this.locale,
       semanticsLabel: this.semanticsLabel,
       textWidthBasis: this.textWidthBasis,
@@ -205,11 +232,14 @@ extension TextStyleExtension on Text {
     );
   }
 
+  /// Sets the decoration property
   Text xDecoration(TextDecoration decoration) {
     return Text(
       this.data,
       key: this.key,
-      style: this.style != null ? this.style.copyWith(decoration: decoration) : TextStyle(decoration: decoration),
+      style: this.style != null
+          ? this.style.copyWith(decoration: decoration)
+          : TextStyle(decoration: decoration),
       locale: this.locale,
       semanticsLabel: this.semanticsLabel,
       textWidthBasis: this.textWidthBasis,
@@ -224,11 +254,14 @@ extension TextStyleExtension on Text {
     );
   }
 
+  /// Sets the decorationThickness property
   Text xDecorationThickness(double decorationThickness) {
     return Text(
       this.data,
       key: this.key,
-      style: this.style != null ? this.style.copyWith(decorationThickness: decorationThickness) : TextStyle(decorationThickness: decorationThickness),
+      style: this.style != null
+          ? this.style.copyWith(decorationThickness: decorationThickness)
+          : TextStyle(decorationThickness: decorationThickness),
       locale: this.locale,
       semanticsLabel: this.semanticsLabel,
       textWidthBasis: this.textWidthBasis,
@@ -243,11 +276,14 @@ extension TextStyleExtension on Text {
     );
   }
 
+  /// Sets the decorationStyle property
   Text xDecorationStyle(TextDecorationStyle decorationStyle) {
     return Text(
       this.data,
       key: this.key,
-      style: this.style != null ? this.style.copyWith(decorationStyle: decorationStyle) : TextStyle(decorationStyle: decorationStyle),
+      style: this.style != null
+          ? this.style.copyWith(decorationStyle: decorationStyle)
+          : TextStyle(decorationStyle: decorationStyle),
       locale: this.locale,
       semanticsLabel: this.semanticsLabel,
       textWidthBasis: this.textWidthBasis,
@@ -262,11 +298,14 @@ extension TextStyleExtension on Text {
     );
   }
 
+  /// Sets the decorationColor property
   Text xDecorationColor(Color decorationColor) {
     return Text(
       this.data,
       key: this.key,
-      style: this.style != null ? this.style.copyWith(decorationColor: decorationColor) : TextStyle(decorationColor: decorationColor),
+      style: this.style != null
+          ? this.style.copyWith(decorationColor: decorationColor)
+          : TextStyle(decorationColor: decorationColor),
       locale: this.locale,
       semanticsLabel: this.semanticsLabel,
       textWidthBasis: this.textWidthBasis,
@@ -281,11 +320,14 @@ extension TextStyleExtension on Text {
     );
   }
 
+  /// Sets the LetterSpacing property
   Text xLetterSpacing(double space) {
     return Text(
       this.data,
       key: this.key,
-      style: this.style != null ? this.style.copyWith(letterSpacing: space) : TextStyle(letterSpacing: space),
+      style: this.style != null
+          ? this.style.copyWith(letterSpacing: space)
+          : TextStyle(letterSpacing: space),
       locale: this.locale,
       semanticsLabel: this.semanticsLabel,
       textWidthBasis: this.textWidthBasis,
@@ -300,11 +342,14 @@ extension TextStyleExtension on Text {
     );
   }
 
+  /// Sets the Height property
   Text xHeight(double space) {
     return Text(
       this.data,
       key: this.key,
-      style: this.style != null ? this.style.copyWith(letterSpacing: space) : TextStyle(letterSpacing: space),
+      style: this.style != null
+          ? this.style.copyWith(letterSpacing: space)
+          : TextStyle(letterSpacing: space),
       locale: this.locale,
       semanticsLabel: this.semanticsLabel,
       textWidthBasis: this.textWidthBasis,
@@ -319,11 +364,14 @@ extension TextStyleExtension on Text {
     );
   }
 
+  /// Sets the inherit boolean property
   Text xInherit(bool inherit) {
     return Text(
       this.data,
       key: this.key,
-      style: this.style != null ? this.style.copyWith(inherit: inherit) : TextStyle(inherit: inherit),
+      style: this.style != null
+          ? this.style.copyWith(inherit: inherit)
+          : TextStyle(inherit: inherit),
       locale: this.locale,
       semanticsLabel: this.semanticsLabel,
       textWidthBasis: this.textWidthBasis,
@@ -338,11 +386,14 @@ extension TextStyleExtension on Text {
     );
   }
 
+  /// Sets the fontFamilyCallback property
   Text xFontFamilyCallback(List<String> callback) {
     return Text(
       this.data,
       key: this.key,
-      style: this.style != null ? this.style.copyWith(fontFamilyFallback: callback) : TextStyle(fontFamilyFallback: callback),
+      style: this.style != null
+          ? this.style.copyWith(fontFamilyFallback: callback)
+          : TextStyle(fontFamilyFallback: callback),
       locale: this.locale,
       semanticsLabel: this.semanticsLabel,
       textWidthBasis: this.textWidthBasis,