+++ CHANGELOG.md
@@ -1,6 +1,26 @@
+## [0.0.1+7] 11/11/2020
+* Usable stable version
+
+## [0.0.1+6] 11/11/2020
+* Added test
+
+## [0.0.1+5] 11/11/2020
+* Refactored Container widget
+
+## [0.0.1+4] 11/11/2020
+* Fixed docs
+
+## [0.0.1+3] 11/11/2020
+* Added basic docs
+
+## [0.0.1+2] 11/11/2020
+
+* Passed flutter pub static analysis
+
 ## [0.0.1+1] 6/11/2020
 
 * Fixed static analysis error
+
 
 ## [0.0.1] 6/11/2020.